+++ src/dendrogram.cpp
@@ -139,4 +139,77 @@ List dendrogram_to_reach(const Rcpp::List x) {
   return(res); 
 }
 
+// [[Rcpp::export]]
+List mst_to_dendrogram(const NumericMatrix mst) {
+
+  // Set up sorted vector values
+  NumericVector p_order = mst(_, 0);
+  NumericVector q_order = mst(_, 1);
+  NumericVector dist = mst(_, 2);
+  int n_nodes = p_order.length() + 1; 
+  
+  // Make sure to clone so as to not make changes by reference
+  p_order = Rcpp::clone(p_order);
+  q_order = Rcpp::clone(q_order);
+
+  // UnionFind data structure for fast agglomerative building
+  UnionFind uf((size_t) n_nodes);
+
+  // Create leaves
+  List dendrogram(n_nodes);
+  for (int i = 0; i < n_nodes; ++i) {
+    IntegerVector leaf = IntegerVector();
+    leaf.push_back(i+1);
+    leaf.attr("label") = patch::to_string(i + 1);
+    leaf.attr("members") = 1;
+    leaf.attr("height") = 0;
+    leaf.attr("leaf") = true;
+    dendrogram.at(i) = leaf;
+  }
+  
+  // Get the index of the point with next smallest reach dist and its neighbor
+  IntegerVector members(n_nodes, 1);
+  int insert = 0, p = 0, q = 0, p_i = 0, q_i = 0;
+  for (int i = 0; i < (n_nodes-1); ++i) {
+    p = p_order(i), q = q_order(i);
+
+    // Get the actual index of the branch(es) containing the p and q
+    p_i = uf.Find(p), q_i = uf.Find(q);
+    
+    // Merge the two, retrieving the new index
+    uf.Union(p_i, q_i);
+    List branch = List::create(dendrogram.at(q_i), dendrogram.at(p_i));
+
+    insert = uf.Find(q_i); // q because q_branch is first in the new branch
+
+    // Update members in the branch
+    int tmp_members = members.at(p_i) + members.at(q_i);
+    
+    // Branches with equivalent distances are merged simultaneously 
+    while((i + 1) < (n_nodes-1) && dist(i + 1) == dist(i)){
+      i += 1;
+      p = p_order(i), q = q_order(i);
+      p_i = uf.Find(p), q_i = uf.Find(q);
+      
+      // Merge the branches, update current insert index
+      int insert2 = uf.Find(q_i); 
+      branch.push_back(insert == insert2 ? dendrogram.at(p_i) : dendrogram.at(q_i)); 
+      tmp_members += insert == insert2 ? members.at(p_i) : members.at(q_i); 
+      uf.Union(p_i, q_i);
+      insert = uf.Find(q_i); 
+
+    }
+    // Generic proxy blocks attr access for mixed types, so need to keep track of members manually!
+    branch.attr("height") = dist(i);
+    branch.attr("class") = "dendrogram";
+    branch.attr("members") = tmp_members;  
+
+    // Update members reference and insert the branch
+    members.at(insert) = branch.attr("members");
+    dendrogram.at(insert) = branch;
+
+  }
+  return(dendrogram.at(insert));
+}
+